Move `ExploreResultsLink` component
