+++ frontend/src/metabase/query_builder/components/view/ViewHeader/ViewHeader.jsx
@@ -18,7 +18,6 @@ import { navigateBackToDashboard } from "metabase/query_builder/actions";
 import { MODAL_TYPES } from "metabase/query_builder/constants";
 import { getDashboard } from "metabase/query_builder/selectors";
 import QuestionActions from "../../QuestionActions";
-import { ExploreResultsLink } from "../ExploreResultsLink";
 import { FilterHeaderButton } from "../FilterHeaderButton";
 import { HeadBreadcrumbs } from "../HeaderBreadcrumbs";
 import QuestionDataSource from "../QuestionDataSource";
@@ -27,6 +26,7 @@ import { QuestionNotebookButton } from "../QuestionNotebookButton";
 import ConvertQueryButton from "../ConvertQueryButton";
 import { FilterHeaderToggle, FilterHeader } from "../QuestionFilters";
 import { QuestionSummarizeWidget } from "../QuestionSummaries";
+import { ExploreResultsLink } from "./components/ExploreResultsLink";
 import {
   AdHocViewHeading,
   SaveButton,
